generalization for authorizeWithRefreshSAS, authorizeWithSAS and authorizeWithSharedKey
