+++ lib/blob.js
@@ -10,6 +10,7 @@ var xml               = require('./xml-parser');
 var util              = require('util');
 var events            = require('events');
 var agent             = require('./agent');
+var auth              = require('./authorization');
 
 /*
  * Azure storage service version
@@ -70,60 +71,6 @@ var QUERY_PARAMS_SUPPORTED = [
   'blockid',
   'blocklisttype'
 ].sort();
-
-/*
- * Authorize the request with shared key
- * Intended to define `Blob.prototype.authorize`.
- */
-var authorizeWithSharedKey = function (method, path, query, headers) {
-  var headersValueToSign = (
-    (headers['content-encoding']     || '') + '\n' +
-    (headers['content-language']     || '') + '\n' +
-    (headers['content-length']       || '') + '\n' +
-    (headers['content-md5']          || '') + '\n' +
-    (headers['content-type']         || '') + '\n' +
-    '\n' + // we always include x-ms-date, so we never specify date header
-    (headers['if-modified-since']    || '') + '\n' +
-    (headers['if-match']             || '') + '\n' +
-    (headers['if-none-match']        || '') + '\n' +
-    (headers['if-unmodified-since']  || '') + '\n' +
-    (headers['range']                || '')
-  );
-  var authOptions = {
-    accountId: this.options.accountId,
-    accessKey: this._accessKey,
-    agent: this.options.agent,
-    hostname: this.hostname,
-    hasCanonicalizedHeaders: true,
-    queryParamsSupported: QUERY_PARAMS_SUPPORTED,
-    headersValueToSign: headersValueToSign
-  };
-
-  return utils.buildRequestOptionsForAuthSharedKey(authOptions, method, path, query, headers);
-}
-
-/*
- * Authorize the request with a shared-access-signature that is refreshed with
- * a function given as `options.sas`.
- * Intended to define `Blob.prototype.authorize`.
- */
-function authorizeWithRefreshSAS(method, path, query, headers) {
-  return utils.buildRequestOptionsForRefreshSAS(this, method, path, query, headers);
-}
-
-/**
- * Authorize the request with a shared-access-signature that is given with
- * `options.sas` as string.
- * Intended to define `Blob.prototype.authorize`.
- */
-function authorizeWithSAS(method, path, query, headers) {
-  var authOptions = {
-    agent: this.options.agent,
-    hostname: this.hostname,
-    sas: this.options.sas
-  };
-  return utils.buildRequestOptionsForSAS(authOptions, method, path, query, headers);
-}
 
 function anonymous(method, path, query, headers) {
   // Serialize query-string
@@ -239,19 +186,19 @@ function Blob(options) {
   // Define `this.authorize`
   if (this.options.accessKey) {
     // If set authorize to use shared key signatures
-    this.authorize = authorizeWithSharedKey;
+    this.authorize = auth.authorizeWithSharedKey.call(this, QUERY_PARAMS_SUPPORTED);
 
     // Decode accessKey
     this._accessKey = new Buffer(this.options.accessKey, 'base64');
   } else if (this.options.sas instanceof Function) {
     // Set authorize to use shared-access-signatures with refresh function
-    this.authorize = authorizeWithRefreshSAS;
+    this.authorize = auth.authorizeWithRefreshSAS;
     // Set state with _nextSASRefresh = -1, we'll refresh on the first request
     this._nextSASRefresh = -1;
     this._sas = '';
   } else if (typeof(this.options.sas) === 'string') {
     // Set authorize to use shared-access-signature as hardcoded
-    this.authorize = authorizeWithSAS;
+    this.authorize = auth.authorizeWithSAS;
   } else {
     this.authorize = anonymous;
   }
