+++ lib/queue.js
@@ -10,6 +10,7 @@ var util        = require('util');
 var agent       = require('./agent');
 var utils       = require('./utils');
 var xml         = require('./xml-parser');
+var auth        = require('./authorization');
 
 /*
  * Azure storage service version
@@ -30,16 +31,6 @@ var TRANSIENT_ERROR_CODES = [
 ].concat(utils.TRANSIENT_HTTP_ERROR_CODES);
 
 /*
- * List of x-ms-* headers supported in lexicographical order, used for
- * construction of the canonicalized headers.
- */
-var X_MS_HEADERS_SUPPORTED = [
-  'x-ms-client-request-id',
-  'x-ms-date',
-  'x-ms-version'
-].sort();
-
-/*
  * List of query-string parameter supported in lexicographical order, used for
  * construction of the canonicalized resource.
  */
@@ -56,171 +47,6 @@ var QUERY_PARAMS_SUPPORTED = [
   'peekonly',
   'popreceipt'
 ].sort();
-
-/*
- * Authorize the request with shared key
- * Intended to define `Queue.prototype.authorize`.
- */
-var authorizeWithSharedKey = function (method, path, query, headers) {
-  // Find account id
-  var accountId = this.options.accountId;
-
-  // Build string to sign
-  var stringToSign = (
-    method + '\n' +
-    (headers['content-encoding']     || '') + '\n' +
-    (headers['content-language']     || '') + '\n' +
-    (headers['content-length']       || '') + '\n' +
-    (headers['content-md5']          || '') + '\n' +
-    (headers['content-type']         || '') + '\n' +
-    '\n' + // we always include x-ms-date, so we never specify date header
-    (headers['if-modified-since']    || '') + '\n' +
-    (headers['if-match']             || '') + '\n' +
-    (headers['if-none-match']        || '') + '\n' +
-    (headers['if-unmodified-since']  || '') + '\n' +
-    (headers['range']                || '')
-  );
-
-  // Check if we have meta-data header fields, if we don't we can use a
-  // presorted list of headers which doesn't involve any allocations.
-  // Otherwise, we callback to building a list of 'x-ms-' prefixed headers and
-  // sorting it.
-  var hasMetadata = false;
-  for (var field in headers) {
-    hasMetadata = hasMetadata || /^x-ms-meta/.test(field);
-  }
-  var fields;
-  if (hasMetadata) {
-    // Construct fields as a sorted list of 'x-ms-' prefixed headers
-    fields = [];
-    for (var field in headers) {
-      if (/^x-ms-/.test(field)) {
-        fields.push(field);
-      }
-    }
-    fields.sort();
-  } else {
-    // Fields used in most API methods presorted in lexicographical order.
-    fields = X_MS_HEADERS_SUPPORTED;
-  }
-
-  // Add lines for canonicalized headers using presorted list of fields
-  var N = fields.length;
-  for(var i = 0; i < N; i++) {
-    var field = fields[i];
-    var value = headers[field];
-    if (value) {
-      stringToSign += '\n' + field + ':' + value;
-    }
-  }
-
-  // Added lines from canonicalized resource and query-string parameters
-  // supported by this library in lexicographical order as presorted in
-  // QUERY_PARAMS_SUPPORTED
-  stringToSign += '\n/' + accountId + path;
-  var M = QUERY_PARAMS_SUPPORTED.length;
-  for(var j = 0; j < M; j++) {
-    var param = QUERY_PARAMS_SUPPORTED[j];
-    var value = query[param];
-    if (value) {
-      stringToSign += '\n' + param + ':' + value;
-    }
-  }
-
-  // Compute signature
-  var signature = utils.hmacSha256(this._accessKey, stringToSign);
-
-  // Set authorization header
-  headers.authorization = 'SharedKey ' + accountId + ':' + signature;
-
-  // Encode query string
-  var qs = querystring.stringify(query);
-
-  // Construct request options
-  return Promise.resolve({
-    host:       this.hostname,
-    method:     method,
-    path:       (qs.length > 0 ? path + '?' + qs : path),
-    headers:    headers,
-    agent:      this.options.agent,
-  });
-}
-
-/*
- * Authorize the request with a shared-access-signature that is refreshed with
- * the a function given as `options.sas`.
- * Intended to define `Queue.prototype.authorize`.
- */
-function authorizeWithRefreshSAS(method, path, query, headers) {
-  var self = this;
-  // Check if we should refresh SAS
-  if (Date.now() > this._nextSASRefresh && this._nextSASRefresh !== 0) {
-    debug("Refreshing shared-access-signature");
-    // Avoid refreshing more than once
-    this._nextSASRefresh = 0;
-    // Refresh SAS
-    this._sas = Promise.resolve(this.options.sas());
-    // Update _nextSASRefresh when the SAS has been refreshed
-    this._sas.then(function(sas) {
-      sas = querystring.parse(sas);
-      // Find next sas refresh time
-      self._nextSASRefresh = (
-        new Date(sas.se).getTime() - self.options.minSASAuthExpiry
-      );
-      debug("Refreshed shared-access-signature, will refresh in %s ms",
-            self._nextSASRefresh);
-      // Throw an error if the signature expiration comes too soon
-      if (Date.now() > self._nextSASRefresh) {
-        throw new Error("Refreshed SAS, but got a Shared-Access-Signature " +
-                        "that expires less than options.minSASAuthExpiry " +
-                        "from now, signature expiry: " + sas.se);
-      }
-    }).catch(function(err) {
-      // If we have an error freshing SAS that's bad and we'll emit it; for most
-      // apps it's probably best to ignore this error and just crash.
-      self.emit('error', err);
-    });
-  }
-
-  // Construct request options, whenever the `_sas` promise is resolved.
-  return this._sas.then(function(sas) {
-    // Serialize query-string
-    var qs = querystring.stringify(query);
-    if (qs.length > 0) {
-      qs += '&';
-    }
-    qs += sas
-    return {
-      host:       self.hostname,
-      method:     method,
-      path:       path + '?' + qs,
-      headers:    headers,
-      agent:      self.options.agent,
-    };
-  });
-}
-
-/*
- * Authorize the request with a shared-access-signature that is given with
- * `options.sas` as string.
- * Intended to define `Queue.prototype.authorize`.
- */
-function authorizeWithSAS(method, path, query, headers) {
-  // Serialize query-string
-  var qs = querystring.stringify(query);
-  if (qs.length > 0) {
-    qs += '&';
-  }
-  qs += this.options.sas;
-  // Construct request options
-  return Promise.resolve({
-    host:       this.hostname,
-    method:     method,
-    path:       path + '?' + qs,
-    headers:    headers,
-    agent:      this.options.agent,
-  });
-}
 
 /**
  * Queue client class for interacting with Azure Queue Storage.
@@ -321,18 +147,18 @@ function Queue(options) {
   // Define `this.authorize`
   if (this.options.accessKey) {
     // If set authorize to use shared key signatures
-    this.authorize = authorizeWithSharedKey;
+    this.authorize = auth.authorizeWithSharedKey.call(this, QUERY_PARAMS_SUPPORTED);
     // Decode accessKey
     this._accessKey = new Buffer(this.options.accessKey, 'base64');
   } else if (this.options.sas instanceof Function) {
     // Set authorize to use shared-access-signatures with refresh function
-    this.authorize = authorizeWithRefreshSAS;
+    this.authorize = auth.authorizeWithRefreshSAS;
     // Set state with _nextSASRefresh = -1, we'll refresh on the first request
     this._nextSASRefresh = -1;
     this._sas = '';
   } else if (typeof(this.options.sas) === 'string') {
     // Set authorize to use shared-access-signature as hardcoded
-    this.authorize = authorizeWithSAS;
+    this.authorize = auth.authorizeWithSAS;
   } else {
     throw new Error("Either options.accessKey, options.sas as function or " +
                     "options.sas as string must be given!");
