+++ lib/table.js
@@ -9,6 +9,7 @@ var events      = require('events');
 var util        = require('util');
 var agent       = require('./agent');
 var utils       = require('./utils');
+var auth        = require('./authorization');
 
 /* Transient error codes (we'll retry request when encountering these codes */
 var TRANSIENT_ERROR_CODES = [
@@ -65,86 +66,6 @@ var authorizeWithSharedKey = function (method, path, query, headers) {
     agent:      this.options.agent,
   });
 }
-
-
-/*
- * Authorize the request with a shared-access-signature that is refreshed with
- * the a function given as `options.sas`.
- * Intended to define `Table.prototype.authorize`.
- */
-function authorizeWithRefreshSAS(method, path, query, headers) {
-  var self = this;
-
-  // Check if we should refresh SAS
-  if (Date.now() > this._nextSASRefresh && this._nextSASRefresh !== 0) {
-    debug("Refreshing shared-access-signature");
-    // Avoid refreshing more than once
-    this._nextSASRefresh = 0;
-    // Refresh SAS
-    this._sas = Promise.resolve(this.options.sas());
-    // Update _nextSASRefresh when the SAS has been refreshed
-    this._sas.then(function(sas) {
-      sas = querystring.parse(sas);
-      // Find next sas refresh time
-      self._nextSASRefresh = (
-        new Date(sas.se).getTime() - self.options.minSASAuthExpiry
-      );
-      debug("Refreshed shared-access-signature, will refresh in %s ms",
-            self._nextSASRefresh);
-      // Throw an error if the signature expiration comes too soon
-      if (Date.now() > self._nextSASRefresh) {
-        throw new Error("Refreshed SAS, but got a Shared-Access-Signature " +
-                        "that expires less than options.minSASAuthExpiry " +
-                        "from now, signature expiry: " + sas.se);
-      }
-    }).catch(function(err) {
-      // If we have an error freshing SAS that's bad and we'll emit it; for most
-      // apps it's probably best to ignore this error and just crash.
-      self.emit('error', err);
-    });
-  }
-
-  // Construct request options, whenever the `_sas` promise is resolved.
-  return this._sas.then(function(sas) {
-    // Serialize query-string
-    var qs = querystring.stringify(query);
-    if (qs.length > 0) {
-      qs += '&';
-    }
-    qs += sas
-    return {
-      host:       self.hostname,
-      method:     method,
-      path:       path + '?' + qs,
-      headers:    headers,
-      agent:      self.options.agent,
-    };
-  });
-}
-
-
-/*
- * Authorize the request with a shared-access-signature that is given with
- * `options.sas` as string.
- * Intended to define `Table.prototype.authorize`.
- */
-function authorizeWithSAS(method, path, query, headers) {
-  // Serialize query-string
-  var qs = querystring.stringify(query);
-  if (qs.length > 0) {
-    qs += '&';
-  }
-  qs += this.options.sas;
-  // Construct request options
-  return Promise.resolve({
-    host:       this.hostname,
-    method:     method,
-    path:       path + '?' + qs,
-    headers:    headers,
-    agent:      this.options.agent,
-  });
-}
-
 
 /**
  * Table client class for interacting with Azure Table Storage.
@@ -263,18 +184,18 @@ function Table(options) {
   // Define `this.authorize`
   if (this.options.accessKey) {
     // If set authorize to use shared key signatures
-    this.authorize = authorizeWithSharedKey;
+    this.authorize = auth.authorizeWithSharedKey.call(this);
     // Decode accessKey
     this._accessKey = new Buffer(this.options.accessKey, 'base64');
   } else if (this.options.sas instanceof Function) {
     // Set authorize to use shared-access-signatures with refresh function
-    this.authorize = authorizeWithRefreshSAS;
+    this.authorize = auth.authorizeWithRefreshSAS;
     // Set state with _nextSASRefresh = -1, we'll refresh on the first request
     this._nextSASRefresh = -1;
     this._sas = '';
   } else if (typeof(this.options.sas) === 'string') {
     // Set authorize to use shared-access-signature as hardcoded
-    this.authorize = authorizeWithSAS;
+    this.authorize = auth.authorizeWithSAS;
   } else {
     throw new Error("Either options.accessKey, options.sas as function or " +
                     "options.sas as string must be given!");
