+++ lib/utils.js
@@ -303,189 +303,6 @@ var isValidGUID = function isValidGUID(value) {
 exports.isValidGUID = isValidGUID;
 
 /**
- * Helper method to build the request options for Shared Key authentication.
- *
- * @param {object} options - Options on the form:
- * ```js
- * {
- *    accountId: '...',                 // Azure storage accountId (required)
- *    accessKey: '...',                 // Decoded Azure shared accessKey (required)
- *    agent: '...',                     // HTTP Agent to use  (required)
- *    hostname: '...',                  // The service hostname (required)
- *    hasCanonicalizedHeaders: false,   // Specify if the string to sign includes the canonicalized headers (optional)
- *    queryParamsSupported: '...'       // List of query-string parameter supported in lexicographical order, used for
- *                                      // construction of the canonicalized resource.
- *    headersValueToSign: '...'         // A string that includes all the headers value separated by a new line
- *                                      // that should be included in the string to sign. (for more details,
- *                                      // see the documentation of every service) (required)
- * }
- * ```
- * @param {string} method - HTTP verb in upper case, e.g. `GET`.
- * @param {string} path - Path on service resource for storage account.
- * @param {object} query - Query-string parameters.
- * @param {object} header - Mapping from header key in lowercase to value.
- *
- * @return {Promise} A promise for an options object compatible with `https.request`.
- */
-var buildRequestOptionsForAuthSharedKey = function buildRequestOptionsForAuthSharedKey(options, method, path, query, headers) {
-  // Find account id
-  var accountId = options.accountId;
-
-  // Build string to sign
-  var stringToSign = (
-    method + '\n' + options.headersValueToSign
-  );
-
-  // Construct fields as a sorted list of 'x-ms-' prefixed headers
-  if (options.hasCanonicalizedHeaders) {
-    var fields = [];
-    for (var field in headers) {
-      if (/^x-ms-/.test(field)) {
-        fields.push(field);
-      }
-    }
-    fields.sort();
-
-    // Add lines for canonicalized headers using presorted list of fields
-    var N = fields.length;
-    for(var i = 0; i < N; i++) {
-      var field = fields[i];
-      var value = headers[field];
-      if (value) {
-        // Convert each HTTP header to lower case and
-        // trim any white space around the colon in the header.
-        stringToSign += '\n' + field.toLowerCase() + ':' + value.trim();
-      }
-    }
-  }
-
-  // Added lines from canonicalized resource and query-string parameters
-  // supported by this library in lexicographical order as presorted in
-  // options.queryParamsSupported
-  stringToSign += '\n/' + accountId + path;
-  var M = options.queryParamsSupported.length;
-  for(var j = 0; j < M; j++) {
-    var param = options.queryParamsSupported[j];
-    var value = query[param];
-    if (value) {
-      stringToSign += '\n' + param + ':' + value;
-    }
-  }
-
-  // Compute signature
-  var signature = hmacSha256(options.accessKey, stringToSign);
-
-  // Set authorization header
-  headers.authorization = 'SharedKey ' + accountId + ':' + signature;
-
-  // Encode query string
-  var qs = querystring.stringify(query);
-
-  // Construct request options
-  return Promise.resolve({
-    host:       options.hostname,
-    method:     method,
-    path:       (qs.length > 0 ? path + '?' + qs : path),
-    headers:    headers,
-    agent:      options.agent
-  });
-}
-
-exports.buildRequestOptionsForAuthSharedKey = buildRequestOptionsForAuthSharedKey;
-
-/**
- * Helper method to build the request options for Shared-Access-Signature authentication.
- *
- * @param {object} options - Options on the form:
- * ```js
- * {
- *    agent: '...',     // HTTP Agent to use  (required)
- *    hostname: '...',  // The service hostname (required)
- *    sas: '...'        // The Shared-Access-Signature string (required)
- * }
- * ```
- * @param {string} method - HTTP verb in upper case, e.g. `GET`.
- * @param {string} path - Path on service resource for storage account.
- * @param {object} query - Query-string parameters.
- * @param {object} header - Mapping from header key in lowercase to value.
- *
- * @return {Promise} A promise for an options object compatible with `https.request`.
- */
-var buildRequestOptionsForSAS = function buildRequestOptionsForSAS(options, method, path, query, headers) {
-  // Serialize query-string
-  var qs = querystring.stringify(query);
-  if (qs.length > 0) {
-    qs += '&';
-  }
-  qs += options.sas;
-  // Construct request optionss
-  return Promise.resolve({
-    host:       options.hostname,
-    method:     method,
-    path:       path + '?' + qs,
-    headers:    headers,
-    agent:      options.agent
-  });
-}
-
-exports.buildRequestOptionsForSAS = buildRequestOptionsForSAS;
-
-/**
- * Helper method to build the request options for Shared-Access-Signature authentication,
- * where the shared-access-signature is refreshed with a function given as `options.sas`
- *
- * @param serviceInstance - `this` service instance
- * @param method - HTTP verb in the upper case, e.g. `GET`
- * @param path - Path on service resource for storage account
- * @param query - Query-string parameters
- * @param headers - Mapping fro header key in lowercase to value
- *
- * @returns {Promise} A promise for an options object compatible with `https.request`
- */
-var buildRequestOptionsForRefreshSAS = function buildRequestOptionsForRefreshSAS(serviceInstance, method, path, query, headers) {
-  // Check if we should refresh SAS
-  if (Date.now() > serviceInstance._nextSASRefresh && serviceInstance._nextSASRefresh !== 0) {
-    debug("Refreshing shared-access-signature");
-    // Avoid refreshing more than once
-    serviceInstance._nextSASRefresh = 0;
-    // Refresh SAS
-    serviceInstance._sas = Promise.resolve(serviceInstance.options.sas());
-    // Update _nextSASRefresh when the SAS has been refreshed
-    serviceInstance._sas.then(function(sas) {
-      sas = querystring.parse(sas);
-      // Find next sas refresh time
-      serviceInstance._nextSASRefresh = (
-        new Date(sas.se).getTime() - serviceInstance.options.minSASAuthExpiry
-      );
-      debug("Refreshed shared-access-signature, will refresh in %s ms",
-        serviceInstance._nextSASRefresh);
-      // Throw an error if the signature expiration comes too soon
-      if (Date.now() > serviceInstance._nextSASRefresh) {
-        throw new Error("Refreshed SAS, but got a Shared-Access-Signature " +
-          "that expires less than options.minSASAuthExpiry " +
-          "from now, signature expiry: " + sas.se);
-      }
-    }).catch(function(err) {
-      // If we have an error freshing SAS that's bad and we'll emit it; for most
-      // apps it's probably best to ignore this error and just crash.
-      serviceInstance.emit('error', err);
-    });
-  }
-
-  // Construct request options, whenever the `_sas` promise is resolved.
-  return serviceInstance._sas.then(function(sas) {
-    var authOptions = {
-      agent: serviceInstance.options.agent,
-      hostname: serviceInstance.hostname,
-      sas: sas
-    };
-    return buildRequestOptionsForSAS(authOptions, method, path, query, headers);
-  });
-}
-
-exports.buildRequestOptionsForRefreshSAS = buildRequestOptionsForRefreshSAS;
-
-/**
  * @param {object} conditionalHeaders - Conditional headers on the following form
  * ```js
  * {
